requirements.txt
# Honeypot Web Application Dependencies
# Production-grade cybersecurity deception platform

# Core Framework
Flask==3.0.0
Werkzeug==3.0.1
gunicorn==21.2.0

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Security & Cryptography
cryptography==41.0.7
PyJWT==2.8.0

# Async & Queue
redis==5.0.1
celery==5.3.4

# Utilities
python-dotenv==1.0.0
bleach==6.1.0
ua-parser==0.18.0

# Rate limiting
Flask-Limiter==3.5.0

# Development (comment out in production)
# flask-debugtoolbar==0.14.1
